requirements.txt
Django~=2.2.4
qrcode
django-simple-history
djangorestframework
Pillow
mysqlclient
